+++ pdfalyzer/helpers/pdf_object_helper.py
@@ -1,5 +1,5 @@
 """
-Some methods to help with the direct manipulation/processing of PyPDF2's PdfObjects
+Some methods to help with the direct manipulation/processing of PyPDF's PdfObjects
 """
 from typing import List, Optional
 
@@ -24,7 +24,7 @@ def _sort_pdf_object_refs(refs: List[PdfObjectRelationship]) -> List[PdfObjectRe
 
 
 def pypdf_class_name(obj: PdfObject) -> str:
-    """Shortened name of type(obj), e.g. PyPDF2.generic._data_structures.ArrayObject becomes Array"""
+    """Shortened name of type(obj), e.g. PyPDF.generic._data_structures.ArrayObject becomes Array"""
     class_pkgs = type(obj).__name__.split('.')
     class_pkgs.reverse()
     return class_pkgs[0].removesuffix('Object')